requirements.txt
Flask ~= 2.0
gunicorn ~= 20.1
flake8 ~= 4.0
requests ~= 2.26
python-dotenv ~= 0.19
PyJWT ~= 2.3
PyJWT[crypto]
black ~= 22.3
sentry_sdk ~= 1.15